move lsmash to its rightful place
