+++ src/vapoursynth/mod.rs
@@ -1,3 +1,4 @@
 pub mod havsfunc;
 #[allow(unused)]
 pub mod output;
+pub mod lsmash;
